// packages/danmaku-anywhere/src/common/danmaku/types/v3/schema.ts
interface BaseDanmakuMeta {
  provider: DanmakuSourceType
}

/**
 * @deprecated
 */
interface DanDanPlayMeta extends BaseDanmakuMeta {
  provider: DanmakuSourceType.DanDanPlay
  /**
   * All properties come from DDP API
   */
  episodeId: number
  animeId: number
  episodeTitle: string
  animeTitle: string
}

/**
 * @deprecated
 */
interface BiliBiliMeta extends BaseDanmakuMeta {
  provider: DanmakuSourceType.Bilibili
  /**
   * All properties come from Bilibili API
   */
  // cid
  cid: number

  // submission id
  bvid?: string
  aid: number

  // season id
  seasonId: number

  title: string
  seasonTitle: string
  mediaType: BiliBiliMediaType
}

/**
 * @deprecated
 */
interface TencentMeta extends BaseDanmakuMeta {
  provider: DanmakuSourceType.Tencent
  cid: string // season id
  vid: string // episode id
  episodeTitle: string
  seasonTitle: string
}

/**
 * @deprecated
 */
interface CustomMeta extends BaseDanmakuMeta {
  provider: DanmakuSourceType.Custom
  seasonTitle: string
  episodeTitle: string
}

/**
 * Danmaku cache is always created by the background script,
 * we don't need to differentiate between request and response types since it's always a response.
 * The only exception is when we're importing danmaku from a file,
 * in which case we assume the file is exported from the extension and has the correct schema,
 * and we can import it without modification.
 *
 * @deprecated
 */
interface BaseEpisodeV3 {
  // The source of the danmaku
  provider: DanmakuSourceType
  comments: CommentEntity[]
  commentCount: number
  // How many times the comments have been updated
  version: number
  // The last time the comments were updated
  timeUpdated: number
  schemaVersion: 1 | 2 | 3

  // Identifier for the episode, varies between providers
  episodeId?: number | string
  episodeTitle: string
  // Season id, used for grouping episodes, can be undefined for some providers
  seasonId?: number | string
  seasonTitle: string

  // Auto generated id
  id: number
}

/**
 * @deprecated
 */
export type DanDanPlayDanmakuV3 = BaseEpisodeV3 & {
  provider: DanmakuSourceType.DanDanPlay
  meta: DanDanPlayMeta
  /**
   * The params used to fetch the comments
   */
  params: Partial<GetCommentQuery>
}

/**
 * @deprecated
 */
export type DanDanPlayDanmakuInsertV3 = Omit<DanDanPlayDanmakuV3, 'id'>

/**
 * @deprecated
 */
export type CustomDanmakuV3 = BaseEpisodeV3 & {
  provider: DanmakuSourceType.Custom
  meta: CustomMeta
}

/**
 * The model used to insert into the database, episodeId is auto generated so is omitted
 * @deprecated
 */
export type CustomDanmakuInsertV3 = Omit<CustomDanmakuV3, 'id'>

/**
 * @deprecated
 */
export type BiliBiliDanmakuV3 = BaseEpisodeV3 & {
  provider: DanmakuSourceType.Bilibili
  meta: BiliBiliMeta
}

/**
 * @deprecated
 */
export type BiliBiliDanmakuInsertV3 = Omit<BiliBiliDanmakuV3, 'id'>

/**
 * @deprecated
 */
export type TencentDanmakuV3 = BaseEpisodeV3 & {
  provider: DanmakuSourceType.Tencent
  meta: TencentMeta
}